requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain_pinecone
langchain
flask
pypdf
python-dotenv
-e .